move get_last_checkpoint fucntion to setup.model
