+++ corebehrt/functional/trainer/setup.py
@@ -1,5 +1,3 @@
-import torch
-
 from corebehrt.modules.setup.config import Config
